requirements.txt
amqp==2.6.0
asgiref==3.2.7
attrs==19.3.0
billiard==3.6.3.0
boto3==1.14.2
botocore==1.17.2
celery==4.4.6
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.7
django-celery-results==1.2.1
django-cors-middleware==1.5.0
django-heroku==0.3.1
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-camel-case==1.2.0
docutils==0.15.2
future==0.18.2
gunicorn==20.0.4
jmespath==0.10.0
kombu==4.6.11
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
psycopg2==2.8.5
psycopg2-binary==2.8.5
py==1.9.0
py-moneyed==0.8.0
pyparsing==2.4.7
pytest==5.4.3
pytest-django==3.9.0
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
wcwidth==0.2.5
whitenoise==5.1.0
